requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
automium==0.2.6
automium-web==0.1.1
beautifulsoup4==4.9.3
blog==1.0.1
boto3==1.17.3
botocore==1.20.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coverage==5.4
distlib==0.3.1
Django==3.1.6
django-filter==2.4.0
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djongo==1.3.3
dnspython==2.1.0
docopt==0.6.2
false==0.1.0
filelock==3.0.12
filter==0.0.0.20200724
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
legacy==0.1.6
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.22.1
Pillow==8.1.0
pyembed==1.3.3
pyembed-markdown==1.1.0
pygame==2.0.1
Pygments==2.7.4
PyJWT==2.0.1
pylint==2.6.0
pymdown-extensions==8.1.1
pymongo==3.11.2
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
soupsieve==2.1
sqlparse==0.2.4
toml==0.10.2
urllib3==1.26.2
virtualenv==20.4.0
virtualenvwrapper-win==1.2.6
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
